fix(router): рефакторинг компонента Applications
Jira issue: VEGA-672
