+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Application';
+export * from './ApplicationRoutes';
 export * from './Error';
 export * from './Loader';
