+++ src/vega-shell.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createBrowserHistory } from 'history';
-import * as singleSpa from 'single-spa';
 
 import { getAppConfig } from '../app-config';
 
@@ -10,25 +9,15 @@ import { Identity } from './utils/identity';
 import { Applications } from './applications';
 import { BrowserMessageBus } from './message-bus';
 
-const HOME_PAGE = '/projects';
 const bus = BrowserMessageBus.create();
 
 const history = createBrowserHistory();
 
 const authMessage = { channel: 'auth', topic: 'login', self: true };
 
-const sendMessageOnAuth = () => {
-  bus.send({ ...authMessage, payload: { loggedIn: true } });
+const handleLoggedInChange = (data: { isLoggedIn: boolean }) => {
+  bus.send({ ...authMessage, payload: { loggedIn: data.isLoggedIn } });
 };
-
-const sendMessageOnLogout = () => {
-  bus.send({ ...authMessage, payload: { loggedIn: false } });
-};
-
-bus.subscribe({ channel: 'auth', topic: 'logged-in' }, () => {
-  // const url = new URL(window.location.href).searchParams.get('redirect-to') ?? HOME_PAGE;
-  // singleSpa.navigateToUrl(url.toString());
-});
 
 const handleGraphqlClientError = (err: ServerError): void => {
   bus.send({ channel: 'error', topic: 'server-error', payload: err, self: true });
@@ -37,8 +26,8 @@ const handleGraphqlClientError = (err: ServerError): void => {
 const { baseApiUrl } = getAppConfig();
 const identity = new Identity({
   apiUrl: `${baseApiUrl}/login`,
-  cbOnAuth: sendMessageOnAuth,
-  cbOnLogout: sendMessageOnLogout,
+  cbOnAuth: () => handleLoggedInChange({ isLoggedIn: true }),
+  cbOnLogout: () => handleLoggedInChange({ isLoggedIn: false }),
 });
 const graphqlClient = createGraphqlClient({
   uri: `${baseApiUrl}/graphql`,
